# Models/SkillLearning.cs
// Copyright (c) 2024 RFull Development
// This source code is managed under the MIT license. See LICENSE in the project root.
namespace ResumeManagementApi.Models
{
    public record class SkillLearning : Learning
    {
    }
}