# Models/SkillLearningListGetResponse.cs
// Copyright (c) 2024 RFull Development
// This source code is managed under the MIT license. See LICENSE in the project root.
using System.Text.Json.Serialization;

namespace ResumeManagementApi.Models
{
    public record class SkillLearningListGetResponse
    {
        [JsonPropertyName("count")]
        [JsonRequired]
        public required long Count { get; init; }
        [JsonPropertyName("learnings")]
        [JsonRequired]
        public required List<SkillLearning> Learnings { get; init; }
    }
}