# schemas/apis/v1.0/openapi.yaml
openapi: 3.0.3
info:
  title: Resume Management API Specification
  version: 1.0.0
  license:
    name: MIT
    url: https://opensource.org/license/mit/
servers:
  - url: http://localhost:5244
    description: Development Server
...
components:
  schemas:
...
    SkillLearning:
      allOf:
        - $ref: '#/components/schemas/Learning'
    SkillLearningUpdate:
      type: object
      properties:
        id:
          type: string
        description:
          type: string
...
    SkillLearningListGetResponse:
      type: object
      required:
        - count
        - learnings
      properties:
        count:
          type: integer
        learnings:
          type: array
          items:
            $ref: '#/components/schemas/SkillLearning'
    SkillLearningListUpdateRequest:
      type: object
      required:
        - learnings
      properties:
        learnings:
          type: array
          items:
            $ref: '#/components/schemas/SkillLearningUpdate'
...
    Learning:
      type: object
      required:
        - id
        - summary
      properties:
        id:
          type: string
        summary:
          type: string
        description:
          type: string
    LearningListGetResponse:
      type: object
      required:
        - totalCount
        - count
        - learnings
      properties:
        totalCount:
          type: integer
          format: int64
        count:
          type: integer
        learnings:
          type: array
          items:
            $ref: '#/components/schemas/Learning'
    LearningGetResponse:
      allOf:
        - $ref: '#/components/schemas/Learning'
    LearningCreateRequest:
      type: object
      required:
        - summary
      properties:
        summary:
          type: string
    LearningCreateResponse:
      type: object
      required:
        - id
      properties:
        id:
          type: string
    LearningUpdateRequest:
      type: object
      properties:
        summary:
          type: string
        description:
          type: string
paths:
...
  /resumes/{id}/skills/{skillId}/learnings:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: skillId
        in: path
        required: true
        schema:
          type: string
    get:
      tags:
        - Resume
      summary: Retrieves a list of learning of a skill.
      operationId: getResumeSkillLearningList
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SkillLearningListGetResponse'
        '400':
          description: Bad Request
        '404':
          description: Not Found
        '503':
          description: Service Unavailable
    put:
      tags:
        - Resume
      summary: Replace the learnings of a skill.
      operationId: replaceResumeSkillLearning
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SkillLearningListUpdateRequest'
      responses:
        '202':
          description: Accepted
        '400':
          description: Bad Request
        '404':
          description: Not Found
        '409':
          description: Conflict
        '503':
          description: Service Unavailable
    delete:
      tags:
        - Resume
      summary: Delete the learnings of a skill.
      operationId: deleteResumeSkillLearning
      responses:
        '202':
          description: Accepted
        '400':
          description: Bad Request
        '404':
          description: Not Found
        '503':
          description: Service Unavailable
...
  /learnings:
    get:
      tags:
        - Learning
      summary: Retrieves a list of learning.
      operationId: getLearningList
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LearningListGetResponse'
        '503':
          description: Service Unavailable
    post:
      tags:
        - Learning
      summary: Create a new learning.
      operationId: createLearning
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LearningCreateRequest'
      responses:
        '201':
          headers:
            Location:
              schema:
                type: string
                format: uri
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LearningCreateResponse'
        '503':
          description: Service Unavailable
  /learnings/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      tags:
        - Learning
      summary: Retrieves a learning.
      operationId: getLearning
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LearningGetResponse'
        '400':
          description: Bad Request
        '404':
          description: Not Found
        '503':
          description: Service Unavailable
    patch:
      tags:
        - Learning
      summary: Update the properties of a learning.
      operationId: updateLearning
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LearningUpdateRequest'
      responses:
        '202':
          description: Accepted
        '400':
          description: Bad Request
        '404':
          description: Not Found
        '409':
          description: Conflict
        '503':
          description: Service Unavailable
    delete:
      tags:
        - Learning
      summary: Delete a learning.
      operationId: deleteLearning
      responses:
        '202':
          description: Accepted
        '404':
          description: Not Found
        '503':
          description: Service Unavailable